stable refresh

+++ program.py
@@ -1,9 +1,8 @@
-# Dharni
-
 import sqlite3 as m
 
-conn=m.connect('test2.db')
-conn2=m.connect('test3.db')
+
+conn=m.connect('gymlog.db')
+conn2=m.connect('foodlog.db')
 conn3=m.connect('fooddetails.db')
 cursor2=conn2.cursor()
 cursor1=conn.cursor()
@@ -11,6 +10,17 @@ cursor3=conn3.cursor()
 import datetime
 
 from prettytable import from_db_cursor
+
+
+def validatenum(data : str):
+    '''
+    Validates a string passed if it only contains nums.
+    '''
+    nums ='1234567890'
+    for i in data:
+        if i not in nums:
+            raise ValueError
+    return data
 
 
 def save():
@@ -28,13 +38,15 @@ def getdate():
     a=input('''Enter workout date number by number starting with day number\nor enter 'today' if you want to log todays workout: ''')
     if a == 'today':      
         y= datetime.datetime.now()
-        return y.strftime('%d/%m/%Y')
-                        
+        return y.strftime('%d/%m/%Y')                   
+    elif a =='done':
+        return a
     else:                   
         try:
             y = int(a)
         except ValueError:
-            print("Enter integers only") #\u001b[31m
+            print("\nEnter integers only") #\u001b[31m
+            
         if type(y) == int:    
             #date='_/_/_'
             date='%s/_/_ ' %y
@@ -48,46 +60,46 @@ def getdate():
             return date          
                 
 
-def logging():
-    seshdate=''               
-        #try:
-    while True:
-        try:
-            seshdate = getdate()        
-            cursor1.execute('''create table '%s' (wname, noofsets, weightsused, repsdone)''' %seshdate)
-        except m.OperationalError:
-            print('Enter a valid input')
-            continue
-        else:
-            break
-    insertingworkout(seshdate)
-            #except:
-                #print('Log for this name already exists. Delete or add to existing log')
+            
 
 def insertingworkout(seshdate):
-    complete=False
-    while not complete:
-        wname=input('''Enter workout name:  or 'done' ''')
-        if wname =='done':
-             complete=True
+    noset=0
+    while True:
+        wname=input('''Enter workout name:  or 'done' ''').title()
+        if wname =='Done':
+             break
         else:
-             noset=int(input('Enter number of sets: '))
-             weightsused=[]
-             repsdone=[]
-             for i in range(noset):
-                 x=input('Enter weight used in set%s: ' %(i+1))
-                 y=input('Enter reps done in set%s: '%(i+1) )
-                 weightsused.append(x)
-                 repsdone.append(y)
+            while True:
+                x=input('Enter number of sets: ')
+                if x =='done':
+                    break
+                else:
+                    try:
+                        noset=int(x)
+                        weightsused=[]
+                        repsdone=[]
+                        for i in range(noset):
+                            while True:
+                                try:
+                                    x=validatenum(input('Enter weight used in set%s: ' %(i+1)))                                  
+                                    y=validatenum(input('Enter reps done in set%s: '%(i+1) ))                                                                                                          
+                                    weightsused.append(x)
+                                    repsdone.append(y)
+                                    break
+                                except:
+                                    print("Enter integers only")
+                            
+                        sweightsused=''
+                        srepsdone=''
+                        for i in weightsused:
+                            sweightsused+='%s kgs '%i
+                        for i in repsdone:    
+                            srepsdone+= '%s reps ' %i                                              
+                        cursor1.execute('''insert into '%s' values('%s', '%s', '%s', '%s')''' %(seshdate, wname, noset, sweightsused, srepsdone))    
+                        break
+                    except:
+                        print('Enter integers only')
             
-             sweightsused=''
-             srepsdone=''
-             for i in weightsused:
-                 sweightsused+='%s kgs '%i
-             for i in repsdone:    
-                 srepsdone+= '%s reps ' %i                       
-    
-             cursor1.execute('''insert into '%s' values('%s', '%s', '%s', '%s')''' %(seshdate, wname, noset, sweightsused, srepsdone))
 
 
 def showtableinfo(x):
@@ -126,22 +138,28 @@ def collectmacros(mealname, Qty):
          return round(r, 1),round(x,1),round(y,1), z                                                                          
 
 def   insertingmeal(mealdate):
-  complete=''
-  while not complete:  
-      tom=input('''Enter time of meal, Breakfast, Lunch, Snacks, Dinner or 'done': ''').title()        
-      if tom == 'Done':
-         complete=True 
-      else:
-          done=''
-          while not done:
-              mealname=input('''Enter mealname or 'done': ''').title()              
-              if mealname=='Done':
-                  done=True
-              else:
-                  Qty=int(input('Enter Qty:  '))
-                  Protein, Carbs, Fat, Calories= collectmacros(mealname, Qty)
-                  cursor2.execute('''insert into '%s' values('%s','%s','%s','%s','%s','%s','%s')''' %(mealdate, mealname, Qty, tom, Protein, Carbs, Fat, Calories))
-                  save2()    
+    complete=''
+    while not complete:  
+        tom=input('''Enter time of meal, Breakfast, Lunch, Snacks, Dinner or 'done': ''').title()        
+        if tom == 'Done':
+            break 
+        else:
+            done=''
+            while not done:
+                mealname=input('''Enter mealname or 'done': ''').title()              
+                if mealname=='Done':
+                    done=True
+                else:
+                    try:                        
+                        Qty=int(input('Enter Qty:  '))
+                        Protein, Carbs, Fat, Calories= collectmacros(mealname, Qty)
+                        cursor2.execute('''insert into '%s' values('%s','%s','%s','%s','%s','%s','%s')''' %(mealdate, mealname, Qty, tom, Protein, Carbs, Fat, Calories))
+                        save2()
+                    except: 
+                        print("Food item may not be in database or make sure you made a valid input")   
+                    
+                
+                       
 
 
 
@@ -154,33 +172,57 @@ while exit != 1:
         command=input('\nEnter 1 to log a NEW workout \nEnter 2 to VIEW previous workout logs \nEnter 3 to DELETE existing log\nEnter 4 to ADD to existing log\nEnter 5 to go BACK: '  )
                     
         if command == '1':
-                        logging()
-                        save()
+            seshdate=''               
+            while True:
+                try:
+                    seshdate = getdate()
+                    if seshdate =='done':
+                        break        
+                    cursor1.execute('''create table '%s' (wname, noofsets, weightsused, repsdone)''' %seshdate)
+                    insertingworkout(seshdate)
+                    showtableinfo(seshdate)
+                    save()
+                except m.OperationalError:
+                    print('\nCheck if a log for this date already exists\nor make sure you made a valid input\n')
+                    continue
+                else:
+                    break
+            
                         
                                                                                                                                                            
         elif command=='2':
             showtables()
             while True:
                 tablename=input('Enter the Session date to view its log: ')
-                try:   
-                    showtableinfo(tablename)
+                if tablename == 'done':
                     break
-                except:
-                    print('No such table exist. Make sure you typed correctly')                               
+                else:
+                    try:   
+                        showtableinfo(tablename)
+                        break
+                    except:
+                        print('No such table exist. Make sure you typed correctly')                               
         
         
         elif command=='3':
                         print('Delete a existing log')
                         showtables()
+                        done=''
                         while True:
                             try:
-                                seshdate = input('Enter date of session you want to delete: ')                        
-                                showtableinfo(seshdate)
-                                break
+                                seshdate = input('Enter date of session you want to delete: ')
+                                if seshdate =='done':
+                                    done=True
+                                    break
+                                else:                        
+                                    showtableinfo(seshdate)
+                                    break
                             except:
                                 print('No such table exist. Make sure you typed correctly')    
-                       
+                        
                         while True:
+                            if done is True:
+                                break
                             confirmation = input('Are you sure you want to delete this log? Y/N: ').lower()
                             if confirmation == 'y':
                                 cursor1.execute('''drop table '%s' ''' %seshdate )
@@ -194,15 +236,19 @@ while exit != 1:
         elif command == '4':
             showtables()
             while True:
-                try:
+                try:                
                     seshdate= input('Enter date of session you want to add to: ')
-                    showtableinfo(seshdate)
-                    break
+                    if seshdate == 'done':
+                        break
+                    else:        
+                        showtableinfo(seshdate)
+                        break
                 except:
-                    print('No such table exist. Make sure you typed correctly')    
-            insertingworkout(seshdate)
-            showtableinfo(seshdate)
-            save()                       
+                    print('No such table exist. Make sure you typed correctly')
+            if seshdate != 'done':            
+                insertingworkout(seshdate)
+                showtableinfo(seshdate)
+                save()                       
         elif command=='5':
             continue
         else:
@@ -214,14 +260,22 @@ while exit != 1:
     
     
     elif command=='2':
-        command=input('\nEnter 1 to proceed with NEW Log \nEnter 2 to ADD to existng log \nEnter 3 to VIEW summary\nEnter 4 to DELETE existing log\nEnter 5 to go BACK  ')
+        command=input('\nEnter 1 to proceed with NEW Log \nEnter 2 to ADD to existng log \nEnter 3 to VIEW summary\nEnter 4 to DELETE existing log\nEnter 5 to go BACK: ')
         if command =='1':
             print('\nNew Log\n')
-            print('_/_/_')
-            mealdate= getdate(input('''Enter date or 'today: '''))
-            cursor2.execute('''create table '%s' (Fname varchar(10), Qty integer, Tom, Protein float, Carbs float, Fat float, Calories integer)''' %mealdate)
-            insertingmeal(mealdate)
-            showtableinfo2(mealdate)
+            while True:
+                try:
+                    mealdate=getdate()
+                    if mealdate=='done':
+                        break                      
+                    cursor2.execute('''create table '%s' (Fname varchar(10), Qty integer, Tom, Protein float, Carbs float, Fat float, Calories integer)''' %mealdate)
+                    insertingmeal(mealdate)
+                    showtableinfo2(mealdate)
+                    break
+                except:
+                    print('\nCheck if a log for this date already exists\nor make sure you made a valid input\n')
+
+            
                                                       
                     
             
@@ -230,64 +284,101 @@ while exit != 1:
         elif command=='2':            
             print('\nAdd to existing log\n')
             showtables2()
-            mealdate = input('Enter date of log you want to add to:  ')
-            showtableinfo2(mealdate)
-            insertingmeal(mealdate)
-            showtableinfo2(mealdate)
+            while True:
+                try:                
+                    mealdate= input('Enter date of log you want to add to:  ')
+                    if mealdate == 'done':
+                        break
+                    else:        
+                        showtableinfo2(mealdate)
+                        break
+                except:
+                    print('No such table exist. Make sure you typed correctly')
+            if mealdate != 'done':            
+                insertingmeal(mealdate)
+                showtableinfo2(mealdate)
+                save2()
             
             
             
         elif command=='3':
             print('View summary')
             showtables2()
-            mealdate=input('Enter the record date:  ')
-            showtableinfo2(mealdate)
-            totalcal=0
-            cursor2.execute('''select Calories from '%s' ;''' %mealdate) 
-            x = cursor2.fetchall()
-            for i in x:
-                totalcal += int(i[0])
-               
+            while True:
+                mealdate=input('Enter the record date:  ')
+                if mealdate == 'done':
+                    break
+                else:
+                    try:   
+                        showtableinfo2(mealdate)
+                        totalcal=0
+                        cursor2.execute('''select Calories from '%s' ;''' %mealdate) 
+                        x = cursor2.fetchall()
+                        for i in x:
+                            totalcal += int(i[0])
+                        
 
-            totalprotein=0
-            cursor2.execute('''select Protein from '%s' ''' %mealdate)
-            x = cursor2.fetchall()
-            for i in x:
-                totalprotein += int(i[0])
+                        totalprotein=0
+                        cursor2.execute('''select Protein from '%s' ''' %mealdate)
+                        x = cursor2.fetchall()
+                        for i in x:
+                            totalprotein += int(i[0])
 
-            totalcarbs=0
-            cursor2.execute('''select Carbs from '%s' ''' %mealdate)
-            x = cursor2.fetchall()
-            for i in x:
-                totalcarbs += int(i[0])
+                        totalcarbs=0
+                        cursor2.execute('''select Carbs from '%s' ''' %mealdate)
+                        x = cursor2.fetchall()
+                        for i in x:
+                            totalcarbs += int(i[0])
 
 
-            totalfat=0
-            cursor2.execute('''select Fat from '%s' ''' %mealdate)
-            x = cursor2.fetchall()
-            for i in x:
-                totalfat += int(i[0])
-            
-            print('Total calories: ', totalcal)
-            print('Total Protein: ', totalprotein)
-            print('Total Carbs: ', totalcarbs)  
-               
-                                                                                                                                                                                                                                                                                
-            proteincal = 4*totalprotein
-            carbscal = 4*totalcarbs
-            fatcal = 9*totalfat
-            others= totalcal - (proteincal+carbscal+fatcal)     
+                        totalfat=0
+                        cursor2.execute('''select Fat from '%s' ''' %mealdate)
+                        x = cursor2.fetchall()
+                        for i in x:
+                            totalfat += int(i[0])
+                        
+                        print('Total calories: ', totalcal)
+                        print('Total Protein: ', totalprotein)
+                        print('Total Carbs: ', totalcarbs)  
+                        
+                                                                                                                                                                                                                                                                                            
+                        proteincal = 4*totalprotein
+                        carbscal = 4*totalcarbs
+                        fatcal = 9*totalfat
+                        others= totalcal - (proteincal+carbscal+fatcal) 
+                        break
+                    except:
+                        print('No such table exist. Make sure you typed correctly')
+                
                                                                                                                                                                                                                                                                                                                             
         elif command =='4':
             showtables2()
-            mealdate=input('Enter date of log to delete:  ')
-            showtableinfo2(mealdate)
-            confirmation= input('Are you sure you want to delete ? Y/N ')
-            if confirmation == 'Y' or 'y':
-                cursor2.execute('''drop table '%s' ''' %mealdate)
-                print("Log deleted succesfully")
-            else:
-                  continue
+            done=''
+            while True:
+                try:
+                    mealdate=input('Enter date of log to delete:  ')
+                    if mealdate =='done':
+                        done=True
+                        break
+                    else:                        
+                        showtableinfo2(mealdate)
+                        break
+                except:
+                    print('No such table exist. Make sure you typed correctly') 
+
+            while True:
+                if done is True:
+                    break
+                confirmation= input('Are you sure you want to delete ? Y/N ').lower() 
+                if confirmation == 'y':
+                    cursor2.execute('''drop table '%s' ''' %mealdate)
+                    print('Log deleted succesfully')
+                    break
+                elif confirmation == 'n':
+                    print("No log deleted")                                 
+                    break
+                else:
+                    print("Enter either y/n") 
                   
         elif command=='5':
                 continue
@@ -296,7 +387,7 @@ while exit != 1:
               
                                          
     elif command=='3':
-        exit =1
+        exit=1
     
     else:
         print('Invalid command \nEnter either 1 or 2')
